=== BEGIN shared-tests === Move shared test from integration_test to shared file
